+++ src/config.ts
@@ -5,7 +5,8 @@ import {
   IntegrationValidationError,
 } from '@jupiterone/integration-sdk-core';
 
-import { createAPIClient } from './client';
+import { APIClient } from './client';
+import { JiraApiVersion } from './jira';
 
 /**
  * A type describing the configuration fields required to execute the
@@ -23,7 +24,15 @@ import { createAPIClient } from './client';
  *
  */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
+  jiraProtocol: {
+    type: 'string',
+    mask: false,
+  },
   jiraHost: {
+    type: 'string',
+    mask: false,
+  },
+  jiraApiVersion: {
     type: 'string',
     mask: false,
   },
@@ -49,8 +58,36 @@ export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
  * Properties provided by the `IntegrationInstance.config`. Normally reflects the
  * same properties defined by `instanceConfigFields`. See note above.
  */
-export interface IntegrationConfig extends IntegrationInstanceConfig {
+export interface JiraIntegrationInstanceConfig
+  extends IntegrationInstanceConfig {
+  /**
+   * The Jira host name and optionally the port and root path. Valid values
+   * include:
+   *
+   * - `"localhost"`
+   * - `"localhost:8080"`
+   * - `"localhost/urlBase"`
+   * - `"example.com"`
+   * - `"subdomain.example.com/urlBase"`
+   *
+   * @see jiraProtocol
+   */
   jiraHost: string;
+
+  /**
+   * An optional configuration of the Jira server host HTTP connection protocol.
+   * Defaults to `'https'`.
+   *
+   * @see jiraHost
+   */
+  jiraProtocol?: string;
+
+  /**
+   * An optional configuration of the Jira server API version. The integration
+   * will attempt to detect the version.
+   */
+  jiraApiVersion?: JiraApiVersion;
+
   jiraUsername: string;
   jiraPassword: string;
 
@@ -77,29 +114,65 @@ export interface IntegrationConfig extends IntegrationInstanceConfig {
   bulkIngestIssues?: boolean;
 }
 
+/**
+ * Properties normalized from the `IntegrationInstance.config` after validation.
+ * This configuration is provided to execution steps.
+ */
+export type IntegrationConfig = JiraIntegrationInstanceConfig & {
+  hostProtocol: 'http' | 'https';
+  hostName: string;
+  hostPort: number;
+  urlBase: string | undefined;
+};
+
 export async function validateInvocation(
-  context: IntegrationExecutionContext<IntegrationConfig>,
+  context: IntegrationExecutionContext<JiraIntegrationInstanceConfig>,
 ) {
   const { config } = context.instance;
 
   if (!config.jiraHost || !config.jiraPassword || !config.jiraUsername) {
     throw new IntegrationValidationError(
-      'config.jiraHost and config.jiraPassword and config.jiraUsername must be provided by the user',
+      'Config requires all of {jiraHost, jiraUsername, jiraPassword}',
     );
   }
 
-  //this regex matches 'localhost', 'example.com', 'subdomain.example.com'
-  //but cannot contain any chars other than letters, numbers, '-', '/' and '.'
-  //the '/' cannot appear at the beginning or end of the string - it's meant to
-  //represent the format 'test.com/jirasubdir'
-  const hostnameRegex =
-    /^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9/][A-Za-z0-9-/]*[A-Za-z0-9])$/;
-  if (!config.jiraHost.match(hostnameRegex)) {
+  if (!isValidJiraHost(config.jiraHost)) {
     throw new IntegrationValidationError(
-      'config.jiraHost must be a valid hostname. (ex: localhost, test.com, sub.test.com)',
+      'jiraHost must be a valid hostname with optional port and root path. (ex: example.com, example.com:2913, example.com/urlBase, subdomain.example.com)',
     );
   }
 
-  const apiClient = createAPIClient(config, context.logger);
+  const normalizedConfig = buildNormalizedInstanceConfig(config);
+  const apiClient = new APIClient(context.logger, normalizedConfig);
   await apiClient.verifyAuthentication();
+
+  context.instance.config = normalizedConfig;
+}
+
+const JIRA_HOST_REGEX =
+  /^((([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9-]*[A-Za-z0-9]))(:(\d{1,4}))?(\/([A-Za-z0-9]+))?$/;
+
+export function isValidJiraHost(jiraHost: string | undefined): boolean {
+  return !!jiraHost && JIRA_HOST_REGEX.test(jiraHost);
+}
+
+export function buildNormalizedInstanceConfig(
+  config: JiraIntegrationInstanceConfig,
+): IntegrationConfig {
+  const match = config.jiraHost.match(JIRA_HOST_REGEX);
+  if (match) {
+    return {
+      ...config,
+      hostProtocol: (config.jiraProtocol as 'http' | 'https') || 'https',
+      hostName: match[1],
+      hostPort: Number(match[6]) || 443,
+      urlBase: match[8],
+    };
+  } else {
+    throw new IntegrationValidationError(
+      `Invalid jiraHost value, could not calculate configuration: ${JSON.stringify(
+        config.jiraHost,
+      )}`,
+    );
+  }
 }