+++ src/jira/index.ts
@@ -1,6 +1,3 @@
-import createJiraClient from "./createJiraClient";
-
-export * from "./types";
-export * from "./JiraClient";
-
-export { createJiraClient };
+export * from './createJiraClient';
+export * from './types';
+export * from './JiraClient';