+++ src/steps/account.ts
@@ -1,43 +1,41 @@
 import {
-    IntegrationStep,
-    IntegrationStepExecutionContext,
-  } from '@jupiterone/integration-sdk-core';
-  
-  import { IntegrationConfig } from '../config';
-  import { createAPIClient } from '../client';
-  import {
-    createAccountEntity,
-  } from "../converters";
-  
-  export const DATA_ACCOUNT_ENTITY = 'DATA_ACCOUNT_ENTITY';
-  
-  export async function fetchAccountDetails({
-    jobState,
-    instance,
-    logger,
-  }: IntegrationStepExecutionContext<IntegrationConfig>) {
-    const config = instance.config;
-    const apiClient = createAPIClient(config, logger);
-    const accountEntity = await jobState.addEntity(
-      createAccountEntity(await apiClient.jira.fetchServerInfo()),
-    );
-  
-    await jobState.setData(DATA_ACCOUNT_ENTITY, accountEntity);
-  }
-  
-  export const accountSteps: IntegrationStep<IntegrationConfig>[] = [
-    {
-      id: 'fetch-account',
-      name: 'Fetch Account Details',
-      entities: [
-        {
-          resourceName: 'Account',
-          _type: 'jira_account',
-          _class: 'Account',
-        },
-      ],
-      relationships: [],
-      dependsOn: [],
-      executionHandler: fetchAccountDetails,
-    },
-  ];
+  IntegrationStep,
+  IntegrationStepExecutionContext,
+} from '@jupiterone/integration-sdk-core';
+
+import { APIClient } from '../client';
+import { IntegrationConfig } from '../config';
+import { createAccountEntity } from '../converters';
+
+export const DATA_ACCOUNT_ENTITY = 'DATA_ACCOUNT_ENTITY';
+
+export async function fetchAccountDetails({
+  jobState,
+  instance,
+  logger,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const config = instance.config;
+  const apiClient = new APIClient(logger, config);
+  const accountEntity = await jobState.addEntity(
+    createAccountEntity(await apiClient.fetchServerInfo()),
+  );
+
+  await jobState.setData(DATA_ACCOUNT_ENTITY, accountEntity);
+}
+
+export const accountSteps: IntegrationStep<IntegrationConfig>[] = [
+  {
+    id: 'fetch-account',
+    name: 'Fetch Account Details',
+    entities: [
+      {
+        resourceName: 'Account',
+        _type: 'jira_account',
+        _class: 'Account',
+      },
+    ],
+    relationships: [],
+    dependsOn: [],
+    executionHandler: fetchAccountDetails,
+  },
+];